修改 将data和game目录变更为对应tar包，以加快解压速度

+++ src/Task.h
@@ -32,7 +32,7 @@ public:
     Config config;
 
 signals:
-    void updateProgress(int now, int count, QString message);
+    void updateProgress(size_t now, size_t count, QString message);
     void failed(QString message);
     void success();
 private:
